+++ host/display/sound_layer.h
@@ -2,6 +2,7 @@
 #define AIRBALL_DISPLAY_SOUND_LAYER_H
 
 #include <cstddef>
+#include <stdint.h>
 
 namespace airball {
 
@@ -12,7 +13,7 @@ public:
 
   size_t period() const;
 
-  virtual void apply(float* buf, size_t frames, size_t pos) const = 0;
+  virtual void apply(int16_t* buf, size_t frames, size_t pos) const = 0;
 
 private:
   const size_t period_;
